style adjustments
